float output

+++ stringato.py
@@ -4,7 +4,7 @@ def extract_floats(string):
 
 def stampa(*args,**kwargs):
 	print("::",*args,**kwargs)
-	
+
 
 def num_word(word,words):
     """Returns the first number after a certain word in a string of words."""
@@ -12,6 +12,6 @@ def num_word(word,words):
     floats = re.compile(f'{word}'+'(\d+\.\d)').findall(words)
     if len(floats[0])==0:
         ints = re.compile(f'{word}'+'(\d+)').findall(words)
-        return int(ints[0])
+        return float(ints[0])
     else:
         return float(floats[0])
